third_party/blink/web_tests/fast/css/getPropertyValue-columns-expected.txt

Bug 111011: getPropertyValue for -webkit-columns returns null, should compute the shorthand value

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS webkitColumnsValue("columns1") is "10px"
PASS webkitColumnsValue("columns2") is "10"
PASS webkitColumnsValue("columns3") is "10px"
PASS webkitColumnsValue("columns4") is "auto"
PASS webkitColumnsValue("columns5") is "2"
PASS webkitColumnsValue("columns6") is "10px 2"
PASS webkitColumnsValue("columns7") is "auto"
NOTE: 'foo' is an illegal CSS value for '-webkit-columns'.
PASS webkitColumnsValue("columns8") is ""
NOTE: If only few longhand properties are specified, getPropertyValue for shorthand property returns empty string.
PASS webkitColumnsValue("columns9") is ""
PASS successfullyParsed is true

TEST COMPLETE
